Fix frequent hang updating ratings
The problem is that emitting Finished signal before calling release() in
MessageReply<MessageType>::SetReply() creates a race condition because
the TagReaderClient connects the Finished() signal to a deleteLater()
slot. This would work except for the fact that SetReply() is called from
a thread (the WorkerPool) that is different from the Object's thread. So
as soon as Finished() is emitted it's possible for the MessageReply
object to be deleted before calling semaphore_.release(). If the
MessageReply is deleted before calling semaphore_.release() then
release() is being called on an invalid object and anything could
happen. On my linux platform it often hangs forever in a futex call.

Moving the emit after semaphore_.release() solves this specific problem
but I'm not sure if this might cause other problems as it's in a core
library.

So this change is more local to the TagReaderClient. Instead of
connecting directly to the deleteLater() slot I connect to a slot
in TagReaderClient() that will call WaitForFinished() before calling()
deleteLater(). WaitForFinished() should never block for any significant
amount of time because either the semaphore will have already been
released or it will shortly be released. It's only used to make sure
that release() has been called on the semaphore before the object is
deleted. It's admittedly hacky. I'm sure there is a better / more
appropriate fix.